+++ src/common/Container/index.tsx
@@ -1,5 +1,4 @@
 import {FC, ReactElement} from 'react'
-import {withRouter} from 'react-router-dom';
 
 import './Container.css'
 
@@ -7,4 +6,4 @@ const Container: FC = ({ children}): ReactElement => (
   <div className='layout-container'>{children}</div>
 )
 
-export default withRouter(Container)
+export default Container