+++ src/common/Header/components/types.ts
@@ -1,5 +1,0 @@
-import {RouteComponentProps} from 'react-router-dom';
-
-import {IRouteParams} from '../../../types/commonTypes';
-
-export type IHeaderProps = RouteComponentProps<IRouteParams>