+++ src/utils/utils.tsx
@@ -1,8 +1,6 @@
-import {match} from 'react-router-dom';
 import {useEffect} from 'react';
 import {cloneDeep, groupBy, sortBy} from 'lodash'
 
-import {IRouteParams} from '../types/commonTypes';
 import {defaultAscensionMaterials, defaultTalentMaterials} from '../charactersData/materials/expandedMaterialInfo';
 import {
   IBooksAndMoraForLevel,
@@ -17,35 +15,10 @@ import {COUNT, MATERIAL, SORTING_INDEX} from '../const/consts';
 import {HEROS_WIT, MORA} from '../charactersData/materials/materialNames';
 import {IAscensionSummary} from '../views/Character/components/AscensionSummary/types';
 
-export const getOrgDataFromMatch = (matchData: match<IRouteParams>): string => {
-  if (matchData?.params?.name) {
-    return `/${matchData.params.name}`
-  }
-
-  return ''
-}
-
-export const createUrl = (urlData: match<IRouteParams> | string, ...rest: Array<string | number>): string => {
-  const org = typeof urlData === 'string' ? `/${urlData}` : getOrgDataFromMatch(urlData)
-
-  return (
-    org +
-    rest.reduce((accumulator: string, currentValue: string | number) => {
-      if (!currentValue) return accumulator
-
-      const newCurrentValue =
-        typeof currentValue !== 'number' && !currentValue.includes('/') && accumulator.slice(-1) !== '/'
-          ? `/${currentValue}`
-          : currentValue
-      return `${accumulator}${newCurrentValue}`
-    }, '')
-  )
-}
-
-export const ScrollToTopOnMount = () => {
+export const ScrollToTopOnMount = ({ name }: { name?: string | undefined }) => {
   useEffect(() => {
     window.scrollTo(0, 0);
-  }, []);
+  }, [name]);
 
   return null;
 }