+++ src/views/Character/components/LevelMaterialsSummary/index.tsx
@@ -1,15 +1,17 @@
 import {FC, ReactElement} from 'react'
-import {withRouter} from 'react-router-dom';
 import {useTranslation} from 'react-i18next';
+import {useParams} from 'react-router-dom';
 
 import {ILevelMaterialsSummaryProps as IProps} from './types';
 import {calculateMaterials} from '../../../../utils/utils';
 import Material from '../Material';
+import {IRouteParams} from '../../../../types/commonTypes';
 
-const LevelMaterialsSummary: FC<IProps> = ({match: {params}, startingLevel, finalLevel}): ReactElement => {
+const LevelMaterialsSummary: FC<IProps> = ({startingLevel, finalLevel}): ReactElement => {
   const {t} = useTranslation(['common', 'material']);
+  const { name } = useParams<IRouteParams>();
 
-  const materialsNeeded = calculateMaterials(params.name, startingLevel, finalLevel)
+  const materialsNeeded = name ? calculateMaterials(name, startingLevel, finalLevel) : []
 
   const isMaterialsSummaryShown = !((startingLevel.lvl === finalLevel.lvl) && (startingLevel.isAscended === finalLevel.isAscended))
 
@@ -29,4 +31,4 @@ const LevelMaterialsSummary: FC<IProps> = ({match: {params}, startingLevel, fina
   )
 }
 
-export default withRouter(LevelMaterialsSummary)
+export default LevelMaterialsSummary