+++ src/views/Character/components/Material/types.ts
@@ -1,8 +1,6 @@
 import {IMaterial} from '../../../../charactersData/types';
 import {WeekDay} from '../../../../const/consts';
 import {Region} from '../../../../charactersData/materials/expandedMaterialInfo';
-import {RouteComponentProps} from 'react-router-dom';
-import {IRouteParams} from '../../../../types/commonTypes';
 
 export interface ITableItemProps {
   data: [string, IMaterial[]]
@@ -23,6 +21,6 @@ export interface IDungeon {
   [key: string]: IDungeonType
 }
 
-export interface ITableItemPropsReworked extends RouteComponentProps<IRouteParams> {
+export interface ITableItemPropsReworked {
   data: [IMaterial[]]
 }