+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+#### 1.1.7 March-2022
+* Changes
+  * Updated react-router to v6
+  * Bug fixes:
+    * Fixed scroll when changing characters in party block (now page scrolls up)
+    * Go back button in the news block was hidden
+
 #### 1.1.6 February-2022
 * Changes
   * Added farming routes for: